requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.6
JPype1-py3==0.5.5.2
konlpy==0.4.4
pkg-resources==0.0.0
PyYAML==3.12
requests==2.18.4
six==1.11.0
slackclient==1.2.1
urllib3==1.22
websocket-client==0.47.0
